poc/requirements.txt
numpy==1.13.3
pyserial==3.4
future
coloredlogs
